requirements.txt
djangorestframework
django==3.2